backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
celery
httpx
structlog
pytest
